# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is gopa-server. Use these APIs to manage different groups, roles of GOPA",
        "title": "Gopa-server Swagger Documentation",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/application/add": {
            "post": {
                "security": [
                    {
                        "Token": []
                    }
                ],
                "description": "Add an application into the database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "application"
                ],
                "summary": "ApplicationAdd",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project",
                        "name": "project",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Role",
                        "name": "role",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "form",
                        "name": "form",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.ApplicationAddForm"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    }
                }
            }
        },
...
        "/api/v1/project/add": {
            "post": {
                "security": [
                    {
                        "Token": []
                    }
                ],
                "description": "Add a project/group into the database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "project"
                ],
                "summary": "ProjectAdd",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project",
                        "name": "project",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Role",
                        "name": "role",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "form",
                        "name": "form",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.ProjectAddForm"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    }
                }
            }
        },
...
        "/api/v1/projectResource/add": {
            "post": {
                "security": [
                    {
                        "Token": []
                    }
                ],
                "description": "Add a group and role to the given resource.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "projectResource"
                ],
                "summary": "ProjectResourceAdd",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project",
                        "name": "project",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Role",
                        "name": "role",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "form",
                        "name": "form",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.ProjectResourceForm"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    }
                }
            }
        },
...
        "/api/v1/projectRole/add": {
            "post": {
                "security": [
                    {
                        "Token": []
                    }
                ],
                "description": "Add a role to the given project into the database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "projectRole"
                ],
                "summary": "ProjectRoleAdd",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project",
                        "name": "project",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Role",
                        "name": "role",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "form",
                        "name": "form",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.ProjectRolesAddForm"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    }
                }
            }
        },
...
        "/api/v1/rego/add": {
            "post": {
                "security": [
                    {
                        "Token": []
                    }
                ],
                "description": "Add a new rego file by given path to the database. If the path already exists, do not insert.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rego"
                ],
                "summary": "RegoAdd",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project",
                        "name": "project",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Role",
                        "name": "role",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "form",
                        "name": "form",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.RegoForm"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    }
                }
            }
        },
...
        "/api/v1/role/add": {
            "post": {
                "security": [
                    {
                        "Token": []
                    }
                ],
                "description": "Add a role into the database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "role"
                ],
                "summary": "RoleAdd",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project",
                        "name": "project",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Role",
                        "name": "role",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "form",
                        "name": "form",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.RoleAddForm"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    }
                }
            }
        },
...
        "/api/v1/user/add": {
            "post": {
                "security": [
                    {
                        "Token": []
                    }
                ],
                "description": "Add a user into the database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "UserAdd",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project",
                        "name": "project",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Role",
                        "name": "role",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "form",
                        "name": "form",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.UserAddForm"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api.ApplicationAddForm": {
            "type": "object",
            "properties": {
                "application_address": {
                    "type": "string"
                },
                "application_name": {
                    "type": "string"
                },
                "application_routers": {
                    "type": "string"
                }
            }
        },
...
        "api.ProjectAddForm": {
            "type": "object",
            "properties": {
                "project_name": {
                    "type": "string"
                }
            }
        },
...
        "api.ProjectResourceForm": {
            "type": "object",
            "properties": {
                "project_name": {
                    "type": "string"
                },
                "resource_name": {
                    "type": "string"
                },
                "role_name": {
                    "type": "string"
                }
            }
        },
        "api.ProjectRolesAddForm": {
            "type": "object",
            "properties": {
                "role_name": {
                    "type": "string"
                },
                "to_project": {
                    "type": "string"
                }
            }
        },
        "api.RegoForm": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "method": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                }
            }
        },
        "api.RoleAddForm": {
            "type": "object",
            "properties": {
                "role_name": {
                    "type": "string"
                }
            }
        },
...
        "api.UserAddForm": {
            "type": "object",
            "properties": {
                "project": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "handler.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "Token": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}